requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
email-validator
pandas
python-docx
pdfplumber
openai
faiss-cpu
anthropic
numpy
openpyxl
python-multipart
reportlab
requests
beautifulsoup4
xlrd
